fix: new initializeState fn - used in 3_main middleware + on newFakeContext
